Nest customer and merchant routes
