+++ web/src/Routes.tsx
@@ -17,21 +17,10 @@
 import React, {lazy, Suspense} from 'react';
 
 import Loading from 'components/common/Loading';
-import CustomerCommitCountProvider from 'components/customer/contexts/CommitCountContext';
-import MerchantProvider from 'components/merchant/contexts/MerchantContext';
+import CustomerRoutes from 'components/customer/Routes';
+import MerchantRoutes from 'components/merchant/Routes';
 import {Switch, Route} from 'react-router-dom';
 
-// Customer Pages
-const CustomerExplorePage = lazy(() => import('components/customer/explore'));
-// Merchant Pages
-const MerchantLandingPage = lazy(() => import('components/merchant/landing'));
-const MerchantSignUpPage = lazy(() => import('components/merchant/sign-up'));
-const MerchantSignInPage = lazy(() => import('components/merchant/sign-in'));
-const MerchantOngoingListingsPage = lazy(() =>
-  import('components/merchant/ongoing-listings')
-);
-// Common Pages
-const ListingDetailsPage = lazy(() => import('components/listing-details'));
 // Design samples
 const DesignSamplesPage = lazy(() => import('components/design-samples'));
 
@@ -39,20 +28,8 @@ const Routes: React.FC = () => (
   <Suspense fallback={<Loading />}>
     <Switch>
       <Route exact path="/design-samples" component={DesignSamplesPage} />
-      <MerchantProvider>
-        <Route exact path="/merchant" component={MerchantLandingPage} />
-        <Route
-          exact
-          path="/merchant/:merchantId"
-          component={MerchantOngoingListingsPage}
-        />
-        <Route exact path="/merchant/sign-up" component={MerchantSignUpPage} />
-        <Route exact path="/merchant/sign-in" component={MerchantSignInPage} />
-      </MerchantProvider>
-      <CustomerCommitCountProvider>
-        <Route exact path="/" component={CustomerExplorePage} />
-        <Route path="/listing/:listingId" component={ListingDetailsPage} />
-      </CustomerCommitCountProvider>
+      <Route path="/merchant/" component={MerchantRoutes} />
+      <Route path="/" component={CustomerRoutes} />
     </Switch>
   </Suspense>
 );
